chore: refactor useGetOpenTrove hook location
